Fix type of HashTable::find as sometimes it wouldn't find the hash even if it exists

+++ std/hash_table.h
@@ -205,7 +205,7 @@ public:
     }
 
     template<typename F>
-    ConstIterator find(unsigned hash, F&& finder) const {
+    ConstIterator find(size_t hash, F&& finder) const {
         if (this->empty()) {
             return this->end();
         }
